+++ src/dmx/hal/uart.c
@@ -49,7 +49,7 @@ static void DMX_ISR_ATTR dmx_uart_isr(void *arg) {
       dmx_head = driver->dmx.head;
       taskEXIT_CRITICAL_ISR(DMX_SPINLOCK(dmx_num));
       if (dmx_head >= 0 && dmx_head < DMX_PACKET_SIZE_MAX) {
-        size_t read_len = DMX_PACKET_SIZE_MAX - dmx_head;
+        int read_len = DMX_PACKET_SIZE_MAX - dmx_head;
         dmx_uart_read_rxfifo(dmx_num, &driver->dmx.data[dmx_head], &read_len);
         dmx_head += read_len;
         taskENTER_CRITICAL_ISR(DMX_SPINLOCK(dmx_num));
@@ -135,7 +135,7 @@ static void DMX_ISR_ATTR dmx_uart_isr(void *arg) {
           }
 
           // Get the delimiter index
-          uint8_t delimiter_idx = 0;
+          int delimiter_idx = 0;
           for (; delimiter_idx <= 7; ++delimiter_idx) {
             const uint8_t slot_value = driver->dmx.data[delimiter_idx];
             if (slot_value != RDM_PREAMBLE) {
@@ -233,7 +233,7 @@ static void DMX_ISR_ATTR dmx_uart_isr(void *arg) {
     // DMX Transmit #####################################################
     else if (intr_flags & DMX_INTR_TX_DATA) {
       // Write data to the UART and clear the interrupt
-      size_t write_len = driver->dmx.size - driver->dmx.head;
+      int write_len = driver->dmx.size - driver->dmx.head;
       dmx_uart_write_txfifo(dmx_num, &driver->dmx.data[driver->dmx.head],
                             &write_len);
       driver->dmx.head += write_len;
@@ -370,7 +370,7 @@ void dmx_uart_set_baud_rate(dmx_port_t dmx_num, uint32_t baud_rate) {
 #endif
 }
 
-void DMX_ISR_ATTR dmx_uart_invert_tx(dmx_port_t dmx_num, uint32_t invert) {
+void DMX_ISR_ATTR dmx_uart_invert_tx(dmx_port_t dmx_num, int invert) {
   struct dmx_uart_t *uart = &dmx_uart_context[dmx_num];
 #if CONFIG_IDF_TARGET_ESP32C6
   uart->dev->conf0_sync.txd_inv = invert;
@@ -415,9 +415,9 @@ uint32_t DMX_ISR_ATTR dmx_uart_get_rxfifo_len(dmx_port_t dmx_num) {
 }
 
 void DMX_ISR_ATTR dmx_uart_read_rxfifo(dmx_port_t dmx_num, uint8_t *buf,
-                                       size_t *size) {
+                                       int *size) {
   struct dmx_uart_t *uart = &dmx_uart_context[dmx_num];
-  const size_t rxfifo_len = uart_ll_get_rxfifo_len(uart->dev);
+  const int rxfifo_len = uart_ll_get_rxfifo_len(uart->dev);
   if (*size > rxfifo_len) {
     *size = rxfifo_len;
   }
@@ -440,9 +440,9 @@ uint32_t DMX_ISR_ATTR dmx_uart_get_txfifo_len(dmx_port_t dmx_num) {
 }
 
 void DMX_ISR_ATTR dmx_uart_write_txfifo(dmx_port_t dmx_num, const void *buf,
-                                        size_t *size) {
+                                        int *size) {
   struct dmx_uart_t *uart = &dmx_uart_context[dmx_num];
-  const size_t txfifo_len = uart_ll_get_txfifo_len(uart->dev);
+  const int txfifo_len = uart_ll_get_txfifo_len(uart->dev);
   if (*size > txfifo_len) *size = txfifo_len;
   uart_ll_write_txfifo(uart->dev, (uint8_t *)buf, *size);
 }
